+++ src/main/java/com/asdflog/blog/config/router/MonitorRouterConfig.java
@@ -1,29 +1,35 @@
-package com.asdflog.blog.config;
+package com.asdflog.blog.config.router;
 
-import com.asdflog.blog.common.handler.MonitorHandler;
-import com.asdflog.blog.sample.SampleHandler;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.CacheControl;
 import org.springframework.http.MediaType;
+import org.springframework.web.reactive.function.BodyInserters;
 import org.springframework.web.reactive.function.server.RequestPredicates;
 import org.springframework.web.reactive.function.server.RouterFunction;
 import org.springframework.web.reactive.function.server.RouterFunctions;
+import org.springframework.web.reactive.function.server.ServerRequest;
 import org.springframework.web.reactive.function.server.ServerResponse;
+import reactor.core.publisher.Mono;
 
 @Configuration
-public class RouterConfig {
-
-  @Bean
-  public RouterFunction<ServerResponse> route(final SampleHandler sampleHandler) {
-    return RouterFunctions
-        .route(RequestPredicates.GET("/hello")
-            .and(RequestPredicates.accept(MediaType.TEXT_PLAIN)), sampleHandler::hello);
-  }
+public class MonitorRouterConfig {
 
   @Bean
   public RouterFunction<ServerResponse> monitorRouter() {
     return RouterFunctions
         .route(RequestPredicates.GET("/monitor/l7check"), MonitorHandler::ok);
+  }
+
+  static class MonitorHandler {
+
+    private static final String OK = "OK";
+
+    public static Mono<ServerResponse> ok(final ServerRequest request) {
+      return ServerResponse.ok().contentType(MediaType.TEXT_PLAIN)
+          .cacheControl(CacheControl.noStore())
+          .body(BodyInserters.fromValue(OK));
+    }
   }
 
 }
