--- src/main/java/com/asdflog/blog/sample/SampleHandler.java
@@ -1,17 +1,0 @@
-package com.asdflog.blog.sample;
-
-import org.springframework.http.MediaType;
-import org.springframework.stereotype.Component;
-import org.springframework.web.reactive.function.BodyInserters;
-import org.springframework.web.reactive.function.server.ServerRequest;
-import org.springframework.web.reactive.function.server.ServerResponse;
-import reactor.core.publisher.Mono;
-
-@Component
-public class SampleHandler {
-
-  public Mono<ServerResponse> hello(final ServerRequest request) {
-    return ServerResponse.ok().contentType(MediaType.TEXT_PLAIN)
-        .body(BodyInserters.fromValue("Hello, bambi!"));
-  }
-}
