+++ src/main/java/com/asdflog/blog/sample/handler/SampleHandler.java
@@ -1,20 +1,17 @@
-package com.asdflog.blog.common.handler;
+package com.asdflog.blog.sample.handler;
 
-import org.springframework.http.CacheControl;
 import org.springframework.http.MediaType;
+import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.BodyInserters;
 import org.springframework.web.reactive.function.server.ServerRequest;
 import org.springframework.web.reactive.function.server.ServerResponse;
 import reactor.core.publisher.Mono;
 
-public class MonitorHandler {
+@Component
+public class SampleHandler {
 
-  private static final String OK = "OK";
-
-  public static Mono<ServerResponse> ok(final ServerRequest request) {
+  public Mono<ServerResponse> hello(final ServerRequest request) {
     return ServerResponse.ok().contentType(MediaType.TEXT_PLAIN)
-        .cacheControl(CacheControl.noStore())
-        .body(BodyInserters.fromValue(OK));
+        .body(BodyInserters.fromValue("Hello, bambi!"));
   }
-
 }
